api/requirements.txt
flask
psycopg2-binary
groq
python-dotenv
gunicorn
flask-cors